+++ json_serializable/lib/src/enum_utils.dart
@@ -33,20 +33,50 @@ String? enumValueMapFromType(
 
     dynamic fieldValue;
     if (annotation == null) {
-      fieldValue = encodedFieldName(jsonEnum.fieldRename, fe.name);
+      if (jsonEnum.valueField != null) {
+        // TODO: fieldRename is pointless here!!! At least log a warning!
+
+        final fieldElementType = fe.type.element2 as EnumElement;
+
+        final e = fieldElementType.getField(jsonEnum.valueField!);
+
+        if (e == null || e.isStatic) {
+          throw InvalidGenerationSourceError(
+            '`JsonEnum.valueField` was set to "${jsonEnum.valueField}", but '
+            'that is not a valid, instance field on '
+            '`${typeToCode(targetType)}`.',
+            element: targetType.element2,
+          );
+        }
+
+        final reader = ConstantReader(fe.computeConstantValue());
+        final valueReader = reader.read(jsonEnum.valueField!);
+        if (valueReader.validValueType) {
+          fieldValue = valueReader.literalValue;
+        } else {
+          throw InvalidGenerationSourceError(
+            '`JsonEnum.valueField` was set to "${jsonEnum.valueField}", but '
+            'that field does not have a type of String, int, or null.',
+            element: targetType.element2,
+          );
+        }
+      } else {
+        fieldValue = encodedFieldName(jsonEnum.fieldRename, fe.name);
+      }
     } else {
       final reader = ConstantReader(annotation);
 
       final valueReader = reader.read('value');
 
-      if (valueReader.isString || valueReader.isNull || valueReader.isInt) {
+      if (valueReader.validValueType) {
         fieldValue = valueReader.literalValue;
       } else {
         final targetTypeCode = typeToCode(targetType);
         throw InvalidGenerationSourceError(
-            'The `JsonValue` annotation on `$targetTypeCode.${fe.name}` does '
-            'not have a value of type String, int, or null.',
-            element: fe);
+          'The `JsonValue` annotation on `$targetTypeCode.${fe.name}` does '
+          'not have a value of type String, int, or null.',
+          element: fe,
+        );
       }
     }
 
@@ -74,10 +104,16 @@ JsonEnum _fromAnnotation(DartObject? dartObject) {
   }
   final reader = ConstantReader(dartObject);
   return JsonEnum(
-      alwaysCreate: reader.read('alwaysCreate').literalValue as bool,
-      fieldRename: enumValueForDartObject(
-        reader.read('fieldRename').objectValue,
-        FieldRename.values,
-        (f) => f.toString().split('.')[1],
-      ));
+    alwaysCreate: reader.read('alwaysCreate').literalValue as bool,
+    fieldRename: enumValueForDartObject(
+      reader.read('fieldRename').objectValue,
+      FieldRename.values,
+      (f) => f.toString().split('.')[1],
+    ),
+    valueField: reader.read('valueField').literalValue as String?,
+  );
+}
+
+extension on ConstantReader {
+  bool get validValueType => isString || isNull || isInt;
 }